fix(scaffolds): widen Go/C solver types to int64 for 64-bit fields
Documented 64-bit fields and results were parsed or returned as Go int (32-bit on GOARCH=386, overflowing valid inputs). Widen main.go, solution.go, and expected-output parsing together across 03, 04, 07, 10, 17, 35, 42, 43, 47, 54, 57; also 57's C Event.id to long long. Stubs still build and their harness parses; only the unimplemented body fails.

+++ 03-easy-max-drawdown/go/main.go
@@ -7,8 +7,8 @@ import (
 )
 
 type input struct {
-	N      int   `json:"n"`
-	Prices []int `json:"prices"`
+	N      int     `json:"n"`
+	Prices []int64 `json:"prices"`
 }
 
 func main() {
