+++ 17-medium-knapsack/go/main.go
@@ -7,8 +7,8 @@ import (
 )
 
 type item struct {
-	Weight int `json:"weight"`
-	Value  int `json:"value"`
+	Weight int   `json:"weight"`
+	Value  int64 `json:"value"`
 }
 
 type input struct {
