+++ 42-hard-fragmented-string-queries/go/solution.go
@@ -1,6 +1,6 @@
 package main
 
-func solve(parts []string, queries [][]int) string {
+func solve(parts []string, queries [][]int64) string {
 	_ = parts
 	_ = queries
 	return ""
