+++ 57-hard-causal-event-replay/go/solution.go
@@ -1,5 +1,5 @@
 package main
 
-func solve(processes int, events []event) []int {
+func solve(processes int, events []event) []int64 {
 	return nil
 }
